backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-WTF==1.1.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1

marshmallow==3.20.1

redis==4.6.0
# rq==1.15.1  # 已移除，改为同步处理

python-dotenv==1.0.0

requests==2.31.0

langchain>=0.2.0
langchain-community>=0.2.0
langchain_openai>=0.1.0
langgraph>=0.1.0
langsmith>=0.1.0
dashscope>=1.14.1

alibabacloud_tea_openapi>=0.3.0
alibabacloud_docmind_api20220711==1.4.7
alibabacloud_credentials>=0.3.0

weaviate-client~=4.5.4

pandas>=1.5.0
numpy>=1.24.0

jieba>=0.42.1
Pillow>=9.5.0

transformers>=4.30.0
torch>=2.0.0

ragas>=0.1.0

black>=23.0.0

pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
factory-boy>=3.3.0

werkzeug>=2.3.0
psutil>=5.9.0

